Almost Done

+++ MTACodersLicence/Models/ResultModels/TestResultModel.cs
@@ -20,7 +20,7 @@ namespace MTACodersLicence.Models.ResultModels
         public string ResultedOutput { get; set; }
 
         [Display(Name = "Execution Time")]
-        public float ExecutionTime { get; set; }
+        public decimal ExecutionTime { get; set; }
 
         [Display(Name = "Memory Used")]
         public float Memory { get; set; }
